+++ libraries/EpoxyPromAvr/src/EEPROM.h
@@ -1,7 +1,0 @@
-#ifndef EPOXY_PROM_AVR_EEPROM_H
-
-#include "EpoxyPromAvr.h"
-#define EEPROM EPOXY_PROM_AVR
-#define EEPROMClass EpoxyPromAvr
-
-#endif