+++ libraries/EpoxyPromEsp/src/EEPROM.h
@@ -1,7 +1,0 @@
-#ifndef EPOXY_PROM_ESP_EEPROM_H
-
-#include "EpoxyPromEsp.h"
-#define EEPROM EPOXY_PROM_ESP
-#define EEPROMClass EpoxyPromEsp
-
-#endif