+++ CHANGELOG.md
@@ -5,6 +5,8 @@
       `digitalWriteFast` (https://github.com/NicksonYap/digitalWriteFast) to
       allow code using these functions (e.g. `digitalWriteFast()`) to compile
       under EpoxyDuino.
+    * Rename `EpoxyPromAvr` to `EpoxyEepromAvr`; `EpoxyPromEsp` to
+      `EpoxyEepromEsp` for consistency and better self-description.
 * 0.6.2 (2021-03-15)
     * Add more stubs for Arduino API functions: `pulseIn()`, `pulseInLong()`,
       `shiftOut()`, `shiftIn()`, I2C and SPI pins (SS, MOSI, MISO, SCK, SDA,